Address message status review feedback
